more websocket facilities and example added
