+++ Gemfile
@@ -16,4 +16,5 @@ gem "rack"
 gem "rack-unreloader"
 gem "roda", "2.26.0"
 gem "sequel"
+gem "rufus-scheduler"
 gem "sqlite3"
