+++ tasks/done.txt
@@ -1,3 +1,5 @@
+- inline video iOS
+- Background video
 - add touch event stop bubbling
 - big bug when trying convert example then windows resize example
 - video callback
